Moved the main function to tsne_main.cpp.
This would allow one to build bhtsne as a library in
addition to a single binary.
