+++ .travis.yml
@@ -1,0 +1,5 @@
+dist: trusty
+language: cpp
+
+script:
+  - g++ sptree.cpp tsne.cpp tsne_main.cpp -o bh_tsne -O2
