+++ Makefile.win
@@ -5,14 +5,17 @@ TARGET = windows
 
 all: $(TARGET) $(TARGET)\bh_tsne.exe
 
-$(TARGET)\bh_tsne.exe: tsne.obj sptree.obj
-	$(CXX) $(CFLAGS) tsne.obj sptree.obj -Fe$(TARGET)\bh_tsne.exe
+$(TARGET)\bh_tsne.exe: tsne_main.obj tsne.obj sptree.obj
+	$(CXX) $(CFLAGS) tsne_main.obj tsne.obj sptree.obj -Fe$(TARGET)\bh_tsne.exe
 
 sptree.obj: sptree.cpp sptree.h
 	$(CXX) $(CFLAGS) -c sptree.cpp
 
 tsne.obj: tsne.cpp tsne.h sptree.h vptree.h
 	$(CXX) $(CFLAGS) -c tsne.cpp
+
+tsne_main.obj: tsne_main.cpp tsne.h sptree.h vptree.h
+	$(CXX) $(CFLAGS) -c tsne_main.cpp
 
 .PHONY: $(TARGET)
 $(TARGET):
